LayoutTests/imported/w3c/web-platform-tests/workers/modules/shared-worker-parse-error-failure-expected.txt


FAIL Module shared worker construction for script with syntax error should dispatch an event named error. assert_equals: expected function "function Event() {
    [native code]
}" but got function "function ErrorEvent() {
    [native code]
}"
FAIL Static import on module shared worker for script with syntax error should dispatch an event named error. assert_equals: expected function "function Event() {
    [native code]
}" but got function "function ErrorEvent() {
    [native code]
}"
